refac: reorganized input formatter structure
